Passed official tests both in diff and valgrind

+++ EX4/StArray.h
@@ -15,8 +15,8 @@ class StArray {
 
 		//class setting methods
 		bool	addStudent(int studentID, char* studentName);
-		bool	addEE_Course(int studentID, int courseNum, char* courseName, int HwNum, int HwWeigh);
-		bool	addCS_Course(int studentID, int courseNum, char* courseName, int HwNum, int HwWeigh, bool HwTakef, char* courseBook);
+		bool	addEE_Course(int studentID, int courseNum, char* courseName, int HwNum, double HwWeigh);
+		bool	addCS_Course(int studentID, int courseNum, char* courseName, int HwNum, double HwWeigh, bool HwTakef, char* courseBook);
 		bool	setHwGrade(int studentID, int courseNum, int HwNum, int grade);
 		bool	setExamGrade(int studentID, int courseNum, int grade);
 		bool	setFactor(int courseNum, int factor);
